# admin-server/docs/swagger.yaml
basePath: /
definitions:
  model.Import:
    properties:
      created_at:
        example: 1682366228
        type: integer
      file_extension:
        example: csv
        type: string
      file_size:
        example: 1024
        type: integer
      file_type:
        example: text/csv
        type: string
      id:
        example: da5554e3-6c87-41b2-9366-5449a2f15b53
        type: string
      importer_id:
        example: 6de452a2-bd1f-4cb3-b29b-0f8a2e3d9353
        type: string
      is_stored:
        example: false
        type: boolean
      metadata:
        $ref: '#/definitions/model.JSONB'
      num_columns:
        example: 8
        type: integer
      num_processed_values:
        example: 128
        type: integer
      num_rows:
        example: 256
        type: integer
      upload_id:
        example: 50ca61e1-f683-4b03-9ec4-4b3adb592bf1
        type: string
      workspace_id:
        example: b2079476-261a-41fe-8019-46eb51c537f7
        type: string
    type: object
  model.Importer:
    properties:
      allowed_domains:
        example:
        - example.com
        items:
          type: string
        type: array
      created_at:
        example: 1682366228
        type: integer
      created_by:
        $ref: '#/definitions/model.User'
      id:
        example: 6de452a2-bd1f-4cb3-b29b-0f8a2e3d9353
        type: string
      name:
        example: Test Importer
        type: string
      template:
        $ref: '#/definitions/model.Template'
      updated_at:
        example: 1682366228
        type: integer
      updated_by:
        $ref: '#/definitions/model.User'
      webhook_url:
        example: https://api.tableflow.com/hooks
        type: string
      workspace_id:
        example: b2079476-261a-41fe-8019-46eb51c537f7
        type: string
    type: object
  model.JSONB:
    additionalProperties: true
    type: object
  model.Template:
    properties:
      created_at:
        example: 1682366228
        type: integer
      created_by:
        $ref: '#/definitions/model.User'
      id:
        example: f0797968-becc-422a-b135-19de1d8c5d46
        type: string
      importer_id:
        example: 6de452a2-bd1f-4cb3-b29b-0f8a2e3d9353
        type: string
      name:
        example: My Template
        type: string
      template_columns:
        items:
          $ref: '#/definitions/model.TemplateColumn'
        type: array
      updated_at:
        example: 1682366228
        type: integer
      updated_by:
        $ref: '#/definitions/model.User'
      workspace_id:
        example: b2079476-261a-41fe-8019-46eb51c537f7
        type: string
    type: object
  model.TemplateColumn:
    properties:
      created_at:
        example: 1682366228
        type: integer
      created_by:
        $ref: '#/definitions/model.User'
      id:
        example: a1ed136d-33ce-4b7e-a7a4-8a5ccfe54cd5
        type: string
      key:
        example: email
        type: string
      name:
        example: Email
        type: string
      required:
        example: false
        type: boolean
      template_id:
        example: f0797968-becc-422a-b135-19de1d8c5d46
        type: string
      updated_at:
        example: 1682366228
        type: integer
      updated_by:
        $ref: '#/definitions/model.User'
    type: object
  model.Upload:
    properties:
      created_at:
        example: 1682366228
        type: integer
      file_extension:
        example: csv
        type: string
      file_name:
        example: example.csv
        type: string
      file_size:
        example: 1024
        type: integer
      file_type:
        example: text/csv
        type: string
      id:
        example: 50ca61e1-f683-4b03-9ec4-4b3adb592bf1
        type: string
      importer_id:
        example: 6de452a2-bd1f-4cb3-b29b-0f8a2e3d9353
        type: string
      is_parsed:
        example: false
        type: boolean
      is_stored:
        example: false
        type: boolean
      metadata:
        example: '{"user_id": 1234}'
        type: string
      num_columns:
        example: 8
        type: integer
      num_rows:
        example: 256
        type: integer
      tus_id:
        example: ee715c254ee61855b465ed61be930487
        type: string
      upload_columns:
        items:
          $ref: '#/definitions/model.UploadColumn'
        type: array
      workspace_id:
        example: b2079476-261a-41fe-8019-46eb51c537f7
        type: string
    type: object
  model.UploadColumn:
    properties:
      id:
        example: 3c79e7fd-1018-4a27-8b86-9cee84221cd8
        type: string
      index:
        example: 0
        type: integer
      name:
        example: Work Email
        type: string
      sample_data:
        example:
        - test@example.com
        items:
          type: string
        type: array
      template_column_id:
        example: a1ed136d-33ce-4b7e-a7a4-8a5ccfe54cd5
        type: string
      upload_id:
        example: 50ca61e1-f683-4b03-9ec4-4b3adb592bf1
        type: string
    type: object
  model.User:
    properties:
      email:
        example: test@example.com
        type: string
      id:
        example: 249ac073-0fe8-4705-82ff-73e3df491b04
        type: string
      recipe:
        example: github
        type: string
      role:
        example: user
        type: string
      time_joined:
        example: 1682366228
        type: integer
    type: object
  types.Res:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  web.ImporterCreateRequest:
    properties:
      name:
        example: Test Importer
        type: string
    type: object
  web.ImporterEditRequest:
    properties:
      allowed_domains:
        example:
        - example.com
        items:
          type: string
        type: array
      name:
        example: Test Importer
        type: string
      webhook_url:
        example: https://example.com/webhook
        type: string
    type: object
  web.TemplateColumnCreateRequest:
    properties:
      key:
        example: first_name
        type: string
      name:
        example: First Name
        type: string
      required:
        example: false
        type: boolean
      template_id:
        example: f0797968-becc-422a-b135-19de1d8c5d46
        type: string
    type: object
host: localhost:3003
info:
  contact:
    email: support@tableflow.com
    name: TableFlow
    url: https://tableflow.com
  description: The backend API of the TableFlow application.
  termsOfService: https://tableflow.com/terms
  title: TableFlow API
  version: "1.2"
paths:
  /admin/v1/import/{id}:
    get:
      description: Get a single import
      parameters:
      - description: Import ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Import'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Get import
      tags:
      - Import
  /admin/v1/importer:
    post:
      description: Create an importer
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.ImporterCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Importer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Create importer
      tags:
      - Importer
  /admin/v1/importer/{id}:
    get:
      description: Get a single importer
      parameters:
      - description: Importer ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Importer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Get importer
      tags:
      - Importer
    post:
      description: Edit an importer
      parameters:
      - description: Importer ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.ImporterEditRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Importer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Edit importer
      tags:
      - Importer
  /admin/v1/importers:
    get:
      description: Get a list of importers
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Importer'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Get importers
      tags:
      - Importer
  /admin/v1/imports:
    get:
      description: Get a list of imports
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Import'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Get imports
      tags:
      - Import
  /admin/v1/settings/api-key:
    get:
      description: Get the current API key
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Get API key
      tags:
      - API Key
    post:
      description: Regenerate the current API key and return the new API key
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Regenerate API key
      tags:
      - API Key
  /admin/v1/template-column:
    post:
      description: Create a template column
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/web.TemplateColumnCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Template'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Create template column
      tags:
      - Template
  /admin/v1/template-column/{id}:
    delete:
      description: Delete a template column
      parameters:
      - description: Template column ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Template'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Delete template column
      tags:
      - Template
  /admin/v1/template/{id}:
    get:
      description: Get a single template
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Template'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Get template
      tags:
      - Template
  /admin/v1/upload/{id}:
    get:
      description: Get a single upload
      parameters:
      - description: Upload ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Upload'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Get upload
      tags:
      - Upload
  /v1/import/{id}:
    get:
      description: Get an individual import
      parameters:
      - description: Import ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Import'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Get import
      tags:
      - External API
  /v1/import/{id}/download:
    get:
      description: Download the import as a file
      parameters:
      - description: Import ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Res'
      summary: Download import
      tags:
      - External API
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"